app/requirements.txt
PyYAML<=4.0.0
dash==1.12.0
dash-bootstrap-components==0.10.1
dash-core-components==1.10.0
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
plotly==4.7.1
numpy==1.18.4
networkx==2.4
Flask>=1.0.0
chatterbot>=1.0.0
chatterbot-corpus>=1.2.0
SQLAlchemy>=1.2
./nativeAPI_wheel/irisnative-1.0.0-cp34-abi3-linux_x86_64.whl